requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
pyyaml==6.0.1
pandas==2.2.0
openpyxl==3.1.2
jinja2==3.1.3
tkinter-tooltip==2.1.0
